cleaned up code and fixed errors
created a Helper file which has a class containing methods to be used everywhere
The helper class currently has a helper method which takes in the information for a movie from the back end and transforms it into information which could be displayed in the movie modal.
Created an image to be displayed when no poster is found. This fixed the error that comes up when trying to access a movie that has no poster.
fixed error that will come up when trying to access a movie without a trailer. replaced by video id 'NA'

Changed the cast to take in a number n of how many cast members would be displayed. fixed error which would come up when not enough cast members are retrieved from the backend.

moved the loading spinner into its own component.

Made general cleanups to the code